Changed the static variables in discrete session to be non-static
The static members in discrete Session are from old design before the
sessions are part of model manager, and they are inconsistent with other
type of sessions. So the logic in discrete Session are modified to make
them non-static. Some operators are also modified to be compatile with
this change.