python: use the right libdir
The libexec directory is *always* wrong for Python packages. It is
supposed to be either lib/ for pure-Python or ${CMAKE_INSTALL_LIBDIR}/
for archful packages. Since SMTK is a mix, everything goes into the
archful directory (Python3 namespace packages, introduced in Python 3.3,
would allow this, but it is not their intended use case).