Extend discrete group op to take a list of entities with new group
This commit extended the group op to take a list of existing entities when creating a new boundary or domain group
in discrete session and add those entities directly to the newly created group if they meet the group member constraints.
This makes it possible for an application to select some entities and create a group with those entities with one step.