Generalize the Exodus session; add a SLAC reader.
This changes the way datasets are referenced in the Exodus session
so that the structure of the reader's output does not have to
match the Exodus reader's convention. Instead of using handles
that refer to specific blocks in an Exodus reader's output,
reference the VTK datasets themselves.

This also adds support for reading SLAC NetCDF files via VTK's reader.

Finally, this fixes several SMTK issues and provides a new operator:
+ Provide a global operator to export model JSON ("export model json").
+ Fix `model::Manager::modelOwningEntity`.
+ Fix problems with the SetProperty operator.
+ Fix the Exodus reader test to prevent the image test from failing.