Redesign the way Collections store associations to a smtk::model.
Previously a smtk::mesh::Collection would only store the association
to a model in a member variable, and not inside the backend interface.
Now with these changes we have added support for the backend interface
to hold the model uuid that we are associated too.

This means that once a collection is stored to disk, loading back the collection
will restore all the proper relationships to the model.