Unset the local version of BUILD_SHARED_LIBS that moab creates.
Certain version of MOAB will create a variable called BUILD_SHARED_LIBS that
will hide the cache version inside SMTK. This mean that if you are building
smtk shared and with a static moab, you will actually build smtk statically.

The fix is to remove the local version of the variable after importing
the moab config file.