MeshOperator now deletes any temporary mesh files that the worker generates.
When a Mesh Worker has finished it currently generates a temporary file
to store the mesh results. This file is now deleted by the MeshOperator, so
that we don't leave around bad mesh files.

In the future all this logic should be handled by remus, once it has
support for remote FileHandles, and remus::JobResult can list multiple files
generated by the worker.