Refactor some logic in json import and export of models and meshes.
When exporting and importing smtk models (.smtk) with meshes in json format, more logics are added so we
have control over whether native models will be saved or loaded when saving and loading smtk models. Also,
when saving smtk models with meshes, a mesh file (.h5m) will be written out for the meshes, and then
loaded back with the model when importing the saved smtk model.