Fixes for toggling visibility of model and meshes.
When a model visibility is changed, if the model is a submodel of another model, we need to label the parent model to be modifed too. Also, when the visibility or color of a mesh being modified, the associated model needs to be labeled modified. This fixes ensures the client side of model manager still have property relationship of all entities and meshes.