smtk::mesh::Collection now has a modified flag.
The modified flag for a mesh collection represents that underlying mesh
database representation has changed such that a write would generate a file
that would not be byte equivalent to the input file.

In the case of mesh databases that are generated in-memory, the modified
flag is set to be true once the first mesh element is added to the database.