Core::coords/adj_iterate now work with non congruent allocations.
adjacencies/coords_iterate previously presumed that all allocations of a given
set would reside inside the same sequence block. When loading multiple files
into the same database this doesn't occur.

The solution to this problem is really simple, all we have to do is
use the min of the ElementSequence when it is smaller than the end iterator
we have been passed.