Add UI change to switch topology vs list view
In this commit, the default tree view behavior is also changed when loading a new model. Now the tree view will expand the model entry in the tree when it is first being loaded to show first hierarchy of entites. The previous behavior is not to expand the new model entry. The idea behind this behavior change is the assumption that when a model is first loaded, the user wants to have a summary view of the model entities.