Fixed a bug where an entity group can't be assigned to "group" item
In a case, where a ModelEntityItem's bitmask is set to be "group", the validation process for the item value will not allow a group with combined entity flags (such as "GROUP_2D | FACE") to be set to it. If the item's bitmask is ONLY specified as group, then we don't have to check further as long as the potential value is a group.