smtk::mesh::PointForEach is more efficient and allows point modification
Previously the implementation of PointForEach would call the virtual functor
method for each individual point. This was really bad for performance. Instead
we now pass a slab of points & coordinates to the PointForEach functor and
it must process them all.

Secondly we want to have the ability for transformations of coordinates. To
handle this use case we pass the coordinates by reference and a boolean
called coordinatesModified. Only if the boolean is updated to true, are
the modified coordinates saved, otherwise they are ignored.