Fix `EntityRef::associations()`.
This was crashing when no associations were present because
it was dereferencing the `end()` iterator.  Don't do that.

Associations made via the attribute (i.e., calling
`attrib->associateEntity(entity)` as opposed to
`entity.associateAttribute(attrib)`) were not properly
completed because the entity's model manager was being eschewed
in favor of the attribute's. Don't do that, either.

Associations recorded in attributes serialized to XML weren't
being added to the corresponding model-manager entities upon
reload because the reader adds entries to ModelEntityItem
instance directly instead of calling `associateEntity()`.
Do that for sure.