Import and export polygon session state.
Now running *"export smtk model"* will generate a file
(which *"import smtk model"* will read) containing the
internal state data for the polygon session.

Note that for the moment, tessellation information is
included to avoid re-running boost-polygon boolean
operations to load files. This may change in the future
as the cost is pretty small even for large-ish models.