Fix model-entity iteration.
Cells that were not free cells of the model were not being exported -- but
only when the free cell did not **directly** reference its bounding cells (as
opposed to indirectly via a k-use-record referencing a shell referencing a
(k-1)-use-record referencing a (k-1)-cell).
Fix this by adding cells whose use-records participate in a bounding
shell (shell, loop, chain).