Model selected signal should be after mesh selected.
When selecting a model entry in the tree view, its meshes are also selected and two signals will be emitted, one for model, one for mesh. Since an application could behave differently on what's selected, we assume selecting the model should make the model active, not the mesh, therefore, emit the model signal after mesh signal.