Fixed the crash when a null collection is parsed for MeshItem.
When using smtErrorMacro to write error messages to logger during serializing attribute system, exceptions will be thrown from Import-ExportJSON, which will cause a Qt Application to crash if the exceptions are not handled, and by default a QApplication does not. To handle exceptions, the application has to implement QApplication::notify().