ENH: Impoved utilization of shared_pointer support
Item, Attribute and Definition now derive from enable_shared_from_this to make it easy to
get the shared pointer from the raw pointer.  This removed the need for pointer().

Updated the code to reflect these changes as well as rmeoving calls to get the shared pointer
that were not needed.