Add a python version of MultiScaleConverter.
This commit adds the python equivalent of the c++ MultiScaleConverter
executable. In the process, it adds Shiboken bindings for VTKDataConverter
and adds an additional constructor to mesh::CellSet and mesh::PointSet to
accept a python-friendly vector and a Moab-friendly set of ids (as opposed to
a mesh::HandleRange, which would require wrapping Moab classes).