Remove auto-generation of SharedPtr.h
This commit is in reference to Issue #40. Originally, the header file
SharedPtr.h was generated to accommodate pre-C++11 builds. Now that we
enforce the use of C++11, the optional use of boost's shared pointers is
unnecessary. This commit removes the mechanisms for choosing between stl's
and boost's shared_ptr, and adds a static header file that uses the stl
version.