Add c'tors to Mesh/Cell/PointSet that accept ConstCollectionPtrs.
Mesh/Cell/PointSets hold a non-const pointer to their parent Collection, but
the creation of these sets is not considered a const action. As a result, new
constructors have been added to Mesh/Cell/PointSet that perform a
const_pointer_cast on a ConstCollectionPtr. This facilitates the const-ifying
of methods in Collection that generate the aforementioned sets.