Warn during compilation if data testing conditions are met.
The requirement for building tests that use test data is the existence of
the data directory AND the explicit use of hdf5. Originally, if the data
directory were specified but hdf5 were not enabled, the tests that use
data were silently dropped. This commit adds a compile-time warning when
the abovementioned condition holds.