Fix polygon split edge bug after model multiblock structure changes.
The problem is that the information keys on children object were not copied when doing shallowcopy of a composite dataset,
and only the metadata stored at composite dataset level is copied (see vtkCompositeDataset::ShallowCopy).
The old code added an information key for ENTITYUUID on the child object, but later on a filter using model-source did
a shallow copy of the output of model-source, which lost the information key of the children. The split arc in polygon
needs that ENTITYID to operate. The fix is to use ChildMetaData stored in composite data level.