Fix the multiblock display test.
The `test2D.json` file contains a model with no child cells;
it is unclear how, with the test as it is, they were ever
being rendered. Fix the test so that it doesn't direct the
multiblock source to use a guess at the model to render.
This allows the source to fall back to the default behavior
of rendering all the tessellated entities present in the manager,
which is what we wanted to begin with.