Move group operator from discete into common
Now we can apply GO to all sessions in SMTK, and if advanced model
is turned on, we can choose to add vertex/edge/face/volume only or
combined. For discrete session, we would still use the old GO which would
transcribe groups between smtk and vtk.
Two test cases are added for discrete and exodus session.