Fix mesh name changing everytime we change visibility
When load a 2dm file twice then change the visibility of the second
mesh in CMB, the mesh name would change everytime correspondingly. It's
because we would set the collectionName to `original_mesh` first and
assign a unique name to mesh everytime (since the current two meshes
would have same name after the reset). An if statement is added so that
we only assign a unique mesh name when it's empty.

Fix smtk issue #10.