BUG: Fix issue using association mask
In a previous commit , the default association rule was set to have a size of 0 (meaning that no model entities were to be asscoiated with it.
Prior to that change the old behaviopr was to create an association rule that was extensible but whose membership mask was set to nothing (which made no sense)
However, the method of setting the association rule via a membership mask made use of the old behavior in that it would just set the mask and assume that the
association was going to be extensible.

Since the main source of using membership masks is through the template reader, the document parser has been changed so that if you specify the association using
a mask, it will also set the asscoaition to be extensible.