In Exodus session, change access path for entity's model UUID.
Originally, an entity's model UUID was accessed through the model interface.
This was causing crashes in cases where the incident entity was not fully in
the model yet. This commit changes the access path to use the session's maps
connecting VTK objects to their representative UUIDs, bypassing the (crashing)
traversal through a model entity's relations.