Fix issue #78 (Edge split in reshape fails).
When an edge is expunged (as happens when the reshaped edge is split),
then the base view calls `updateUI()` so that the view can remove
the deleted entity from any drop-down menus. However, in the case of
the arc panel, the arc widget was being deleted. Since the arc widget
object was the source of the operation's execution, this was causing
memory corruption and crashes. Don't do that.