Fix deserialization of Exodus session colors...
... and other model-entity properties.

This should close #84.

The Exodus session I/O delegate ignored the contents of the JSON file
except for a sequential list of UUIDs to assign and the list of URLs to read.
Now it imports property records from the JSON file when asked to load the
native model; and if the native models are **not** being loaded, it imports
all of the JSON data.