Lazy initialize Moab IncrementalAllocator.
Originally, the incremental allocator was allocating an initial number
of points upon construction. This resulted in the first 64 points in
each Moab database being reserved, regardless of whether or not the
incremental allocator was used. This commit makes the incremental
allocator initialize lazily, so points are only reserved when the
incremental allocator is requested.