Add assign color to ENTITY_LIST feature
Now user can assign color to ENTITY_LIST by just clicking the icon.
For this ENTITY_LIST, it's parent entity must be a model and the color
info is stored as a FloatProperty on the associated model as "$Entity::
flagSummary(bitflags) + color."(Ex. "face color").
This MR also fix a logic error in qtItemDelegate when deciding color
is mutable or not.