Add the ability to attach datasets to meshes.
Introduce smtk::mesh::DataSet, which provides the ability to assign to
the cells and points of meshsets double-valued data of user-defined
tuple length. Change InterpolateMeshElevation operator to
InterpolateMesh and change its functionality from modifying the
z-coordinate of a mesh's points to adding interpolated data to a
meshset's cells or points according to the mesh's centroid. Augment
ImportVTKData andExportVTKData to accept and create double-valued VTK
cell data, respectively.