BUG: Fix logic error in qtCheckItemComboBox and qtModelEntityItem
In order to be able to pick multiple items in comboBox, CMB uses
a QStandardItemModel in comboBox. But CMB does not update the model
when comboBox inits. Now CMB would initialize the model every time
and disable the pickability of the entity summary.

When entities have been expunged, qtModelEntityItem should also expunge
the corresponding items. A logic is added to address this issue.