Reshape: Do not always promote point 0.
The "tweak edge" operator's "promote" item (listing indices
of points along an edge to promote to model vertices) was
using the invalid entry created by default because the item
used the default setting for `NumberOfRequiredValues` --
which recently (~2-3 months) changed from 0 to 1.

This was causing another bug to appear: reshaping multiple
times failed to re-tessellate the face.