Remove ItemDefs and AttDefs.
Added API to remove ItemDefinitions in attribute::Definition and
GroupItemDefinition. Only removal of local ItemDefinitions supported.

Added API to remove Definitions from attribute::System. Only those
Definitions that are not derived from are allowed to be removed. If
a user needs to remove a base Definitions, it needs to remove its
children types first.

Delete AttDef button is disabled if an AttDef with derived types is
currently selected.