Fix bad URL rewrite.
The Exodus session should not attempt to rewrite the URL
string property on each model; that is now done (in a
less error-prone manner) by `SaveJSON::prepareToSave()`.
The code this commit removes was rewriting an already-relative
path to be relative using a method that did not check whether
the new URL existed or not.

This commit also includes logic to properly turn the Exodus
session's superfluous "modelFiles" JSON entry containing relative
paths into absolute paths based on the SMTK URL.