Update embedded python's module search heuristics.
Using boost::dll, this commit updates PythonInterpreter's logic to
attempt to find the smtk module by keying off the location of the
smtkCore library. Methods have been added to facilitate adding to the
embedded interpreter's PYTHONPATH for additional modules. Also, python
operator logic has been updated to safely allow for the possibility of
a module to not be found. This update requires boost 1.64.0 or newer.