ENH: Improving Model and Mesh Entity Selection in Qt
1. If there is only one entity that matches the selection critera it is now
automatically selected
2. If the selection criteria is that there is only one item to be selected then
if a new item is selected, the previous one is becomes unselected
3. If there are model entities initially selected the model item will attempt to be
set to the pre-selected entities.
4. qtModelEntityItem::associateEntities now pre-validates the list of model entities
being associated by on the following criteria:

a. Filter the list based on the model item's definition to remove inappropriate entities
b. Based on the number of valid entities compare that to the model item's definition size
requirements