Make models a resource.
+ Convert model Manager to inherit Resource.
+ Convert model Entity to inherit ResourceComponent.

Because ResourceComponents are managed as shared pointers,
this means the model manager's map from UUIDs to Entity objects
turns into a map from UUIDs to shared pointers.

Also, ResourceComponents store their UUID.
This is fine but much of the model manager API assumes the UUID
is stored separately (as the key to the map from UUIDs to Entity
instances). This commit makes some preliminary changes so things
compile but the API could be further cleaned up to deal with
Entity creation (with or without a UUID specified).