BUG: Fixing OperatorView Issue concerning duplicate attributes
If teh name Name attribute of the operator view did not exactly match the
name of the attribute used by the operator, another attribute was created.

This change now sets the Name attribute when the operator GUI is being constructed.  Hence it
is no longer required that the Operator View have a Name attribute specified.