BUG: Fixed issue related to Selection Manager and qtModelItem
The problem was that by default every time a qtModelItem was displayed it would
update itself based on the curent selection.  While this behavior is correct for
Operators it is not correct for defining a Simulation.

A mode has been added to qtViews, qtItems and qtAttribute to indicate if this
behavior is desirable - note that the class can still choose to ignore this option and
use the older manual mode - for example if an attrbute had two model enity items that
shared the same selection masks