Add a pqProxy for the PV SMTK model representation.
Hook up the representation's SMTK selection manager so that
when the selection is changed a redraw is requested.

This proxy is also intended to subsume the duties of
vtkSMSMTKModelRepresentationProxy and to inform (via a
request to the SMTK resource manager) the server-side
representation of the server-side SMTK selection to use
for rendering.

This class should eventually move to
`smtk/extension/paraview/appcomponents` but cannot until the other
classes move (to avoid cyclic dependencies).