BUG: Fix the check logic in the qtCheckItemComboBox
When the model entity item is not extensible and its required value
is set to one, qtCheckItem would not handle the selection properly.
Though the text shows that the new item is selected, actually it's
empty. This MR fix this bug by selecting new item after the removal
of the previous item.