BUG: Delete qt views when close data or CMB
SMTK operators use a struct to store pointers to Operator,
qtUIManager, QFrame and qtBaseView derived class. It's a shabby design
because it makes memory management difficult and error-prone. This
MR fix a bug that in the past we never delete custom view instances
and another one that for operators which listen to item modified signal,
they would crash CMB if all data has been closed.