smtk::resource::Lock: add "DoNotLock" state.
In order for operations to call operations on a resource passed
into the top-level operation, there needs to be a mechanism for
skipping resource locking. This commit adds lock state
"DoNotLock", which bypasses resource locking. It is now the default
state for both ResourceItems and ComponentItems. Also, if the lock
state is "DoNotLock" then no value is transcribed to the attribute
file.