Develop registration process for managers and plugins.
This commit introduces smtk::common::Registry, a reference counted object
that manages the lifetime of the registration of a logical collection of
smtk resources & operations (usually denoted by its own namespace) and
instances of managers. It also introduces the singleton PluginManager for
centralizing plugin registration. Finally, it removes the concept of
environments that hold static managers and automatic registration.