FindPVPythonEnvironment: check for pvpython-launcher
On some operating systems and in certain configurations, ParaView creates a
launcher for each of its executables. The launcher is responsible for
setting up the appropriate environment and then launching the real
executable. When a launcher is not available, it is likely because the
rpath for the real executable has been set to accomplish the task of
configuring the environment. We first check if ParaView has a launcher for
pvpython, and fall back to using the real pvpython.