Check for pvpython in $ParaView_CLEXECUTABLES_DIR before target
SMTK has a CMake routine to find pvpython and use it to construct
the right environment for smtk python. When on Windows and when
building SMTK in developer mode, the paraview found by SMTK is
its build tree. This is fine for building, but the build tree
pvpython does not work on Windows. We therefore export the path
to the install tree version of pvpython so SMTK can first check
for this variable before getting pvpython through its paraview
target.