Introduce smtkDefaultPlugins library
With the new smtk plugin macro, whenever an smtk macro is created
inside of smtk its plugin target is cached. After the core library
and all of its sessions are configured, a library is generated whose
sole function is to link against the plugins and expose them to a
paraview-derived app.