Fix client-server logic to get legacy pv readers to work.
Prior to this fix, the cmb geometry reader was not loaded properly
into a consuming program, and attempts to use it would result in a
segfault. The consuming program would also segfault when it encountered
an unknown file extension, since it would attempt to instantiate each
reader to query its validity for the file type.