Encapsulate resource generating paraview pipeline sources.
SMTK resources are represented in ParaView pipelines as the result of a
vtkSMTKSource algorithm. Internally, vtkSMTKSource contains a resource
generating class (vtkSMTKResourceGenerator) whose job is to construct a
resource and create a vtkMultiBlockDataSet from the resource. The resulting
resource is then manipulated in situ by SMTK operations; these operations
flag the source (vtkSMTKSource) for update. By encapsulating the resource
generator within vtkSMTKSource, we can trigger an execution of the visualizaion
pipeline without affecting a regeneration of the resource. The resource will
be regenerated only if the exposed methods of the resource generator (e.g.
FileName) are modified. This design is based off of ParaView's vtkMetaReader.