Refactor the interface between paraview and smtk resources
This commit introduces vtkSMTKResource, a server-side pipeline source that
represents an SMTK resource. It also appends pqSMTKBehavior to construct a
vtkSMTKResource when the pipeline source for a resource is queried but cannot
be found. As a result, any operation that results in the construction of a
new resource will now trigger the rendering of the resource (in addition to
the paraview-style readers and resource creation actions).