Only render 1 copy of the data at a time.
The model representation previously never turned off
the visibility of "normal"ly-rendered blocks when an
object was selected; it merely turned on the "selection"
actor's copy of the block.

Now, at most 1 of the 2 will be drawn (0 will be drawn
if the block is hidden).
Opacity is not used for highlighted/selected blocks.
Note that hidden blocks will not be highlighted at all.
This may be counterintuitive.