BUG: Fixed issue with assign color operation
- Custom View was creating its own copy of the operation instead of using the
one passed to it.  Since it was using a weak pointer to hold onto it, the operation
was deleted when it went out of scope.  It now uses the operation passed to it via
the operation view info object.

- Custom View now holds onto the operation via a shared pointer to match how the
basic qtOperationView's approach

- Custom View was still using the old ActiveViewModel approach that is no longer
supported.  It has been updated not to do so.

- view::ComponentPhraseModel is now created using a view::View::Component instead
of a view::View.  This will allow it's behavior to be customizable in the future (via
the info provided by the qtComponentItem class).