Make editing model-entity names work.
This does not affect attribute (or meshset names), although those
should now be un-editable (i.e., double-clicking on the name will
have no effect rather than popping up an editor that is ignored).

This also does some work to make colors editable, but does not
complete it; the render window is not updated properly.