Fix a problem with incorrect managers being used.
An operation's specification may have a resource manager
but it should never be used by a qtUIManager displaying an
operation's specification since that will rarely if ever
be tied to resources managed by a GUI.

Instead, have the operation panel provide the proper
resource manager by holding on to the pqSMTKWrapper object
created when attaching to a pqServer. That way, the items
the UI manager configures will know what resources to monitor
for changes.