smtk::mesh::ExtractTessellation: use hash map for point id lookup
The method smtk::mesh::PointSet::find is extremely slow, and should be
used only as a last resort. This commit replaces a call to find()
within a tight loop with the construction and repeated query of a hash
map.

NOTE: with this update, the mesh session now loads large datasets as
quickly as the VTK session.