Aim to make mesh + mesh-set visibility toggle-able.
The mesh representation (currently a vtkSMTKModelRepresentation) can
now be toggled on/off properly.

However, block visibilities are toggled but not reflected in the render
view since the vtkMeshMultiBlockSource does not annotate its output
with UUIDs the same way the model source does.

There's probably also some code in vtkSMTKModelRepresentation that
specifically looks for an upstream source of the wrong type.