Make selection render style a user preference.
At this point it is an application-wide user preference.
It might also be exposed as a per-representation property
in the Property Inspector panel, but it feels more like an
application setting than something users want to adjust all
the time.

Also, it could be implemented to mirror the `Representation`
StringVectorProperty in the SMTKModelRepresentation proxy
(see `smconfig.xml`) to more closely match the ParaView pattern.