Fix bad Qt index created handling operation results.
The problem was that `std::copy(...)` was writing data into
a vector of integers specifying the path to an item in the
descriptive phrase hierarchy, but not actually inserting
values. The final addition to the parent's path would then
be the only entry in the path rather than the tail entry.