Add linux option to force the use of old dtags
On some *nix distributions, RPath has been deprecated in favor for RunPath.
One difference between these concepts is that RunPath is not inherited. This
causes issues with locating libraries without manually setting
LD_LIBRARY_PATH when the indirect dependency is in a nonstandard location.
This flag switches the compiler to use RPath in favor of RunPath.