Fix query operation for qtModelEntityAttributeView.
Do not attempt to decode the search string into a bit vector
and back again; if the conversion back to a string had been
done with `flagToSpecifierString()` instead of `flagSummary()`,
it would have worked for auxiliary surfaces, but instead we
should just avoid converting at all; that will make the view
proof against future changes to the query string format.

This also adds `*` as a valid model-entity specifier string
(to mean the same as `any`).