BUG: Fixing IsUnique Logic
The main problem was that all Definitions by default were marked isUnique.  This has now been changed to not be the case.  Note this did not effect saved attribute files since the value of isUnique was always being saved and never assumed.

This change exposed a bug in attribute::Resource findUniqueBaseDefinition method which would always return the definition passed into it as Unique (even when it wasn't) as long as it didn't have a Base Definition.  This has been fixed

This commit also removed the exchange mechanism on the qt Associate Widget (since no one was using it and it prevented the ability to do extended selection).  Extended selection on both the Current and Available List have now been added.